+++ NEWS.md
@@ -1,3 +1,5 @@
+# covid19ita 0.6.0
+
 * added impact plots for Veneto region
 * removed first page regional intensive care